support NONE and NOSASL basic auth type
